:tada: Add while loop usecase #137
